Refactor Cypress test and command set up
